+++ src/classes/router/ws.ts
@@ -8,7 +8,8 @@ export default class RouteWS<GlobContext extends Record<any, any> = {}, Context 
 	/** Generate WS Endpoint */
 	constructor(
 		/** The Path of the Routes */ path: Path | RegExp,
-		/** The Validations to add */ validations: RoutedValidation[] = []
+		/** The Validations to add */ validations: RoutedValidation[] = [],
+		/** The Headers to add */ headers: Record<string, Buffer> = {}
 	) {
 		if (isRegExp(path)) {
 			this.data = {
@@ -17,11 +18,9 @@ export default class RouteWS<GlobContext extends Record<any, any> = {}, Context 
 				path,
 				pathStartWith: '/',
 				data: {
-					validations
-				}, context: {
-					data: {},
-					keep: true
-				}
+					validations,
+					headers
+				}, context: { data: {}, keep: true }
 			}
 		} else {
 			this.data = {
@@ -30,17 +29,21 @@ export default class RouteWS<GlobContext extends Record<any, any> = {}, Context 
 				path,
 				pathArray: path.split('/'),
 				data: {
-					validations
-				}, context: {
-					data: {},
-					keep: true
-				}
+					validations,
+					headers
+				}, context: { data: {}, keep: true }
 			}	
 		}
 	}
 
 	/**
 	 * Add a default State for the Request Context (which stays for the entire requests / websockets lifecycle)
+	 * 
+	 * This will set the default context for the request. This applies to all callbacks
+	 * attached to this handler. When `keepForever` is enabled, the context will be shared
+	 * between requests to this callback and therefore will be globally mutable. This may be
+	 * useful for something like a message or request counter so you dont have to worry about
+	 * transferring it around.
 	 * @example
 	 * ```
 	 * const controller = new Server({ })
@@ -71,14 +74,23 @@ export default class RouteWS<GlobContext extends Record<any, any> = {}, Context 
 	): this {
 		const keepForever = options?.keepForever ?? false
 
-		this.data.context.data = context
-		this.data.context.keep = keepForever
+		this.data.context = {
+			data: context,
+			keep: keepForever
+		}
 
 		return this
 	}
 
 	/**
-	 * Add an Event when someone wants to Connect to the Socket
+	 * Attach a Callback for when someone wants to Upgrade from HTTP a Socket
+	 * 
+	 * This will attach a callback for when the client sends an http request but
+	 * wants to upgrade to a websocket connection. This callback will probably only
+	 * ever be used to initialize the context or validate the request before transferring
+	 * to a connection. when wanting to stop the server from upgrading, call the 2nd
+	 * function argument to force ending the connection process and switching to a normal
+	 * http request.
 	 * @example
 	 * ```
 	 * const controller = new Server({ })
@@ -101,7 +113,11 @@ export default class RouteWS<GlobContext extends Record<any, any> = {}, Context 
 	}
 
 	/**
-	 * Add an Event when someone Establishes a connection to the Socket
+	 * Attach a Callback for when someone Establishes a connection to the Socket
+	 * 
+	 * This will attach a callback for when the websocket connection is established
+	 * and ready for use. This callback will commonly be used to attach references using
+	 * `.printRef()` or similar.
 	 * @example
 	 * ```
 	 * const controller = new Server({ })
@@ -124,7 +140,11 @@ export default class RouteWS<GlobContext extends Record<any, any> = {}, Context 
 	}
 
 	/**
-	 * Add an Event when someone sends a Message to the Socket
+	 * Attach a Callback for when someone sends a Message to the Socket
+	 * 
+	 * This will attach a callback for when the server recieves a message from
+	 * the Client. This event can be disabled completely by setting `message.enabled`
+	 * to false in the initial Server Options.
 	 * @example
 	 * ```
 	 * const controller = new Server({ })
@@ -147,7 +167,12 @@ export default class RouteWS<GlobContext extends Record<any, any> = {}, Context 
 	}
 
 	/**
-	 * Add an Event when someone Disconnects from the Socket
+	 * Attach a Callback for when the Socket closes
+	 * 
+	 * This will attach a callback for when the Socket is closed in any way
+	 * that should trigger this (basically all). In this callback `.message`
+	 * will be the message the client sent as reason for the close or just empty
+	 * when the client didnt send a reason or the server closed the socket.
 	 * @example
 	 * ```
 	 * const controller = new Server({ })